+++ src/main/java/absolutelyaya/ultracraft/damage/DamageTypeTags.java
@@ -18,4 +18,5 @@ public class DamageTypeTags
 	public static final TagKey<DamageType> EXPLODE_PLUSHIE = TagKey.of(RegistryKeys.DAMAGE_TYPE, new Identifier(Ultracraft.MOD_ID, "explode_plushie"));
 	public static final TagKey<DamageType> BREAK_MAGNET = TagKey.of(RegistryKeys.DAMAGE_TYPE, new Identifier(Ultracraft.MOD_ID, "break_magnet"));
 	public static final TagKey<DamageType> PUNCH = TagKey.of(RegistryKeys.DAMAGE_TYPE, new Identifier(Ultracraft.MOD_ID, "punch"));
+	public static final TagKey<DamageType> REDUCED_KNOCKBACK = TagKey.of(RegistryKeys.DAMAGE_TYPE, new Identifier(Ultracraft.MOD_ID, "reduced_knockback"));
 }